review: renaming "util" directory to "utils" for consistency
