+++ src/serlio/PRTContext.cpp
@@ -19,7 +19,7 @@
 
 #include "PRTContext.h"
 
-#include "util/LogHandler.h"
+#include "utils/LogHandler.h"
 
 #include <mutex>
