+++ src/serlio/modifiers/MayaCallbacks.h
@@ -21,8 +21,8 @@
 
 #include "encoder/IMayaCallbacks.h"
 
-#include "util/LogHandler.h"
-#include "util/Utilities.h"
+#include "utils/LogHandler.h"
+#include "utils/Utilities.h"
 
 #include "maya/MObject.h"
