+++ src/serlio/modifiers/PRTModifierAction.cpp
@@ -22,9 +22,9 @@
 #include "modifiers/PRTModifierCommand.h"
 #include "modifiers/RuleAttributes.h"
 
-#include "util/LogHandler.h"
-#include "util/MayaUtilities.h"
-#include "util/Utilities.h"
+#include "utils/LogHandler.h"
+#include "utils/MayaUtilities.h"
+#include "utils/Utilities.h"
 
 #include "prt/StringUtils.h"
