+++ src/serlio/modifiers/RuleAttributes.cpp
@@ -19,8 +19,8 @@
 
 #include "modifiers/RuleAttributes.h"
 
-#include "util/LogHandler.h"
-#include "util/Utilities.h"
+#include "utils/LogHandler.h"
+#include "utils/Utilities.h"
 
 #include "prt/Annotation.h"
