+++ src/serlio/utils/MArrayWrapper.h
@@ -19,7 +19,7 @@
 
 #pragma once
 
-#include "util/MArrayIteratorTraits.h"
+#include "utils/MArrayIteratorTraits.h"
 
 #include <iterator>
 #include <type_traits>
