+++ src/serlio/utils/ResolveMapCache.h
@@ -19,7 +19,7 @@
 
 #pragma once
 
-#include "util/Utilities.h"
+#include "utils/Utilities.h"
 
 #include <chrono>
 #include <map>
